requirements.txt
streamlit
groq
python-dotenv
sympy
plotly
numpy